Solution: Better Naming
